pyproject.toml
[tool.poetry]
name = "interviewprep"
version = "0.1.0"
description = ""
authors = ["Mike Jones <mwj12@case.edu>"]

[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.dev-dependencies]
ipdb = "^0.12.0"
pytest = "^5.0"
sphinx = "^2.1"
pyhamcrest = "^1.9"
isort = "^4.3"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
pytest-mock = "^1.10"
pre-commit = "^1.17"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
